In EditImageFragment set seek bars parameters seekbar_brightness seekbar_constraint seekbar_saturation

+++ app/src/main/java/com/example/shareimageapp/PhotoEditor/Interface/EditImageFragmentListener.java
@@ -2,8 +2,8 @@ package com.example.shareimageapp.PhotoEditor.Interface;
 
 public interface EditImageFragmentListener {
     void onBrightnessChanged(int brightness);
-    void onSaturationChanged(int saturation);
-    void onContrastChanged(int contrast);
+    void onSaturationChanged(float saturation);
+    void onContrastChanged(float contrast);
     void onEditStarted();
     void onEditCompleted();
 }
